requirements.txt
streamlit
plotly
numpy
numpy-stl